Added delete option to occupancy calendar, besides that small bugs are fixed with scale level boundaries for min earned points

+++ FishingBookerLibrary/FishingBookerLibrary/BusinessLogic/MoneyFlowCRUD.cs
@@ -18,7 +18,7 @@ namespace FishingBookerLibrary.BusinessLogic
             return SSMSDataAccess.LoadData<MoneyFlow>(sql).FirstOrDefault();
         }
 
-        public static int UpdatePercentage(decimal percentage)
+        public static int UpdatePercentage(int percentage)
         {
 
             MoneyFlow data = new MoneyFlow
